requirements.txt
astroid==2.2.5
beautifulsoup4==4.8.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.3.1
python-dotenv==0.10.3
rauth==0.7.3
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
typed-ast==1.4.0
urllib3==1.25.5
werkzeug==0.16.0
wrapt==1.11.2
